Change `admin` key to `shutdown` key
The admin key has been changed to shutdown as this reflects what is
actually being typed on the CLI and displayed in the config.
Also this key has been changed to a boolean type as there are only two
states an interface can take.
By default the interface is disabled, this respects security best
practices.