Change `admin` subkey of `neighbor` to `shutdown`
The subkey of `neighbor` has been changed to shutdown as this reflects what is
actually being typed on the CLI and displayed in the config.
Also this key has been changed to a boolean type as there are only two
states this can take.
By default the state of the bgp neighbor is down so it need to be
explicitely be brought up.